+++ src/App.jsx
@@ -13,6 +13,7 @@ import { VendorStoreProvider } from '@/lib/vendorStore'
 import { ADMIN_ROLES, ADMIN_LAYOUT_ROLES, PORTAL_ROLES, SAP_PORTAL_ROLES, SUPER_ADMIN_ROLES, BOARD_ROLES, ENGAGEMENT_ROLES, SETTINGS_ROLES, PRESIDENT_ROLES, FINANCE_ROLES } from '@/lib/permissions'
 import ProtectedRoute from '@/components/auth/ProtectedRoute'
 import BetaTermsAckGate from '@/components/BetaTermsAckGate'
+import { PageHeaderProvider } from '@/lib/pageHeader'
 import AppLayout from '@/components/layout/AppLayout'
 import DashboardPage from '@/pages/DashboardPage'
 import CalendarPage from '@/pages/CalendarPage'
@@ -100,6 +101,7 @@ function App() {
             <ForumStoreProvider>
             <VendorStoreProvider>
             <BrowserRouter>
+              <PageHeaderProvider>
               <BetaTermsAckGate>
               <Routes>
               {/* Public */}
@@ -246,6 +248,7 @@ function App() {
               <Route path="*" element={<Navigate to="/" replace />} />
               </Routes>
               </BetaTermsAckGate>
+              </PageHeaderProvider>
             </BrowserRouter>
             </VendorStoreProvider>
             </ForumStoreProvider>